Semi-fixed typeDeclaration in parser
Something wrong with how tokens were scanned, so when it sought the CLASS, it found something else instead.
This has not been fixed as such, but the order in which tokens are sought, is flipped, so now, if it's not
an interface, it will at least get the class token right. Toby/Tufan